run install with deps (#2729)
Trying to do this in a project: `"postinstall":
"PLAYWRIGHT_BROWSERS_PATH=0 npx playwright install --with-deps
chromium"`

But it fails because --with-deps need to install os-level deps and can't
do that under the `moose` user.

It seems that the monorepo build already installs as root, so this
changes the non-monorepo builds to do the same.